feat : 자체 제작 테이블
