+++ src/components/flight/FlightView.tsx
@@ -2,7 +2,7 @@ import { flightResultData } from 'common/store/atom'
 import { FlightList } from 'common/type/FlightType'
 import DataGridViewer from 'components/common/Not use/DataGridViewer'
 import ScreenTitle from 'components/common/ScreenTitle'
-import TableViewer from 'components/common/TableViewer'
+import TableViewer from 'components/tableView/TableViewer'
 import React, { useTransition } from 'react'
 import { useRecoilValueLoadable } from 'recoil'
 import NavCloseButton from '../navbar/NavCloseButton'
