+++ src/components/tableView/TableViewer.tsx
@@ -1,10 +1,10 @@
 import { FlightList } from 'common/type/FlightType'
-import React, { useState } from 'react'
+import React, { useEffect, useState } from 'react'
 import ViewerRow from './ViewerRow'
 import styled from '@emotion/styled'
 
 const ViewerContainer = styled.div`
-    
+    width:100%;
 `
 
 interface Props {
@@ -12,10 +12,12 @@ interface Props {
 }
 function TableViewer({data} : Props) {
     const [tableData, setTableData] = useState(data ? data : null);
+
+
   return (
     <ViewerContainer>
-        {data?.data?.map(t => {
-            return <ViewerRow data={t}/>
+        {tableData?.data?.map(t => {
+            return <ViewerRow data={t} pk={t.id} key={t.id}/>
         })}
 
     </ViewerContainer>
