+++ src/components/tableView/ViewerColumn.tsx
@@ -1,0 +1,17 @@
+import React from 'react'
+
+interface ColumnType {
+    type?: string
+    pk: number,
+    data?: number | string | null,
+}
+
+function ViewerColumn({ data, type }: ColumnType) {
+    return (
+        <div>
+            {data}
+        </div>
+    )
+}
+
+export default ViewerColumn
